Clean data generated by react-scripts
react-scripts has non-configurable output directory based on src
location. Because it is not present under maven target directory, it
also mean it is not cleaned when invoking clean lifecycle.

Add build (and node_modules) to the list of directories to be removed by
maven-clean-plugin and also use resources section to get react app
copied into jar instead of manually configuring maven-resource-plugin.