Fix flakey Apprunner error
Fix Quarkus apprunner mojo failing from to time when trying to start the
Quarkus application with an argument mismatch error when trying to set
the default exit handler.

The issue is caused by the introduction of another method with the same
name in Quarkus ApplicationLifecycleManager, and the mojo not guarding
against/checking the method arguments.

Fix addresses this by explicitly looking for a method accepting a
BiConsumer argument.

Fixes projectnessie/nessie#357