Fix build issue on OSX, when net.alchim31.maven:scala-maven-plugin confuses the classpath
On OSX it (may?) happen that the Nessie build fails with errors like this:

```
[ERROR] Failed to execute goal net.alchim31.maven:scala-maven-plugin:4.4.0:compile (scala-compile-first) on project nessie-deltalake-spark2: Execution scala-compile-first of goal net.alchim31.maven:scala-maven-plugin:4.4.0:compile failed: A required class was missing while executing net.alchim31.maven:scala-maven-plugin:4.4.0:compile: scala/Function1
...
[ERROR] urls[31] = file:/Users/snazy/.m2/repository/org/scala-lang/scala-library/2.12.10/scala-library-2.12.10.jar
[ERROR] urls[57] = file:/Users/snazy/.m2/repository/org/scala-lang/scala-reflect/2.12.8/scala-reflect-2.12.8.jar
...
[ERROR]
[ERROR] -----------------------------------------------------
[ERROR] : scala.Function1
```

I.e. despite it should use 2.12.12 everywhere, the classpath for the plugin/compiler gets messed up. Just putting in `<scalaVersion>2.12.12</scalaVersion>` into the plugin configuration didn't help, but adding an explicit `<dependency>` to the plugin works. I left the `<scalaVersion>` configuration, because IIRC that's the one passed to zinc.