Introduce `resetStoreUnsafe` operation on new `AdminApi` interface.
PR #576 introduced a function to clear the version-store for tests before each test runs.
The the only test class using this was `TestRest` by using `@Inject` to get the version-store implementation.
That however fails with `@NativeImageTest`, because native-tests don't support `@Inject`.

So the idea here is to introduce a REST functionality to programmatically reset version-store
for test and development purposes. As none of the existing interfaces (`ConfigApi`, `TreeApi`, `ContentsApi`)
are good fits for such an "reset-unsafe" operation, this PR introduces a new `AdminApi` interface.

The `AdminApi.resetStoreUnsafe()` functionality is disabled by default and needs to be explicitly
enabled in the configuration. Later on, it should probably also get some special permissions, but that's
out of scope here.