Replace checkstyle with Spotless & Google Java Style
Formats `*.java`, `*.scala` and `*.xml` files.

Does not format markdown or typescript or python or CSS or HTML files.
For typescript: it feels easier to add some tslint/eslint into the node build process.
For python: it feels easier to add (actually: leave) it to the python linter (or flake8 or whatever).
CSS/HTML: there's not much for format right now, and Eclipse WTP fails to format typescript-ish CSS files.

Spotless can use Prettier, but it does an `npm install` for every module, starts a prettier server and shuts it down - takes quite long.