DRAFT: JDBC store implementation
Still a draft, but passes all integration tests.

The implementation supports multiple dialects to support different databases. At the moment, hsqldb, H2, PostgresQL, CockroachDB and Oracle are implemented.

The code in this PR is based on #700 and definitely not ready for a review, but a proof-of-concept.

All values end in a single row (no relations in the database schema). Databases must support the SQL arrays (or a similar mechanism).

Trees (aka the `children` or an `L1` and `Ref`) are mapped as 43 distinct columns to allow rather simple SQL conditions and updates.
Branch-commits in `Ref` work for up to 15 elements. No explicit overflow-protection implemented, it would probably just throw a lot of exceptions, if that limit is exceeded.

Integration tests run against an in-JVM, on-heap hsqldb + h2 instance by default. Integration tests can also run against PostgresQL, CockroachDB and Oracle (see TESTING.md).

Quarkus native-image tests have been adopted to test INMEMORY + JDBC (HSQL in-memory).

The "Nessie-Lambda" doesn't include JDBC.