WIP: Jackson-based Store worker
The current store worker implementation used with the quarkus server
requires each content type to have some equivalent protobuf schema in
order to be serialized back and forth between JSON <-> POJO <-> Proto.

Replace the handwritten protobuf schemas with jackson protobuf module,
which analyze the POJO object to generate the object on the fly, and
leverage jackson to directly write the protobuf binary representation of
the object.