Add RandomSource utility class
In order for some unit tests to be repeteable, it is required to be able
to set the seed for the random generator used by the source.

Currently it requires code to accept a Random instance argument with the
test framework to set the seed before hand, but this approach makes code
heavier (the Random instance needs to be passed between function calls)
for what should is usually an internal concern.

Replace this pattern with a new utility class named `RandomSource` which
behave exactly like `ThreadLocalRandom` except that it is possible to
temporarily override the current random generator with one initialized
with a specific seed.