Paging for get-commit-log (#935)
Implements paging for the `TreeApi.getCommitLog()`, also updates only the API for `TreeApi.getEntries()`.

Paging for the commit log is implemented purely on the REST endpoint. Since `VersionStore.getCommits()`
already returns a `Stream`, we can just `Stream.limit()` the result. The implementation is quite dumb,
it requests "one more" entry from the version-store and if it sees that "one more" entry, it sets
`hasMore=true` and `token=hash-of-that-one-more-thing`, so the next page starts at the given token.

Generally speaking, clients and users of the API _must not_ interpret the token in any way, but just
treat it as some opaque thing. The API docs now include a sentence that the structure/format of the
token can change without prior notice.

`TreeApi.getEntries()` got the same two new parameters that `TreeApi.getCommitLog()` got, just to make
the API consistent. There is no paging implementation for the entries/keys yet.