Fix handling of unsaved commits in getUpdateState() (#915)
From `TieredVersionStore.commit`, in the `store.load` for current+expected, `TieredVersionStore.ensureValidL1` is called, which calls `InternalBranch.getUpdateState`, which performs a loop over the `unsavedCommits`: it's `lastL1` is initialized to `lastSavedL1` and `lastL1` is "moved iteratively forward in history" via `lastL1.getChildWithTree(c.commit, tree, c.keyMutationList).withCheckpointAsNecessary(store)`.
That `withCheckpointAsNecessary()` occasionally throws a `(Reference)NotFoundException` when it tries to load an unsaved L1 via the `HistoryRetriever`, causing the commit-operation to fail.

As @jacques-n pointed out: The process is this: Imagine we have
* L1.0 (saved)
* L1.1 (unsaved)
* L1.2 (unsaved)
When we start, we create an updatestate for L1.1. In that we try to retrieve L1.0 which saved and succeed.
The problem comes when we then step to L1.2. In L1.2 the `HistoryRetriever` attempts to load L1.1 but we're currently still in the `getUpdateState` loop and so L1.1 is pending save but hasn't actually been save. This is by design, we want to batch up all the L1s that need to be saved and save them at once. The problem is that the `HistoryRetriever` is going direct to the `Store` rather than also seeing that the current "transaction" does have the L1.1 value. We need to rework this path so that the `HistoryRetriever` can also access the in-flight L1.1 value and not fail.


The explicit test fails with the following exception, if the line
```
unsavedL1s.put(lastL1.getId(), lastL1);
```
is commented out in `org.projectnessie.versioned.impl.InternalBranch#getUpdateState`. That change would effectively make the prod-code change a no-op.

```
org.projectnessie.versioned.store.NotFoundException: Unable to find requested ref.
	at org.projectnessie.versioned.dynamodb.DynamoStore.load(DynamoStore.java:228)
	at org.projectnessie.versioned.impl.HistoryRetriever$HistoryIterator.calculateNextList(HistoryRetriever.java:200)
	at org.projectnessie.versioned.impl.HistoryRetriever$HistoryIterator.computeNext(HistoryRetriever.java:136)
	at org.projectnessie.versioned.impl.HistoryRetriever$HistoryIterator.computeNext(HistoryRetriever.java:110)
	at com.google.common.collect.AbstractIterator.tryToComputeNext(AbstractIterator.java:141)
	at com.google.common.collect.AbstractIterator.hasNext(AbstractIterator.java:136)
	at java.base/java.util.Iterator.forEachRemaining(Iterator.java:132)
	at java.base/java.util.Spliterators$IteratorSpliterator.forEachRemaining(Spliterators.java:1801)
	at java.base/java.util.stream.AbstractPipeline.copyInto(AbstractPipeline.java:484)
	at java.base/java.util.stream.AbstractPipeline.wrapAndCopyInto(AbstractPipeline.java:474)
	at java.base/java.util.stream.ReduceOps$ReduceOp.evaluateSequential(ReduceOps.java:913)
	at java.base/java.util.stream.AbstractPipeline.evaluate(AbstractPipeline.java:234)
	at java.base/java.util.stream.ReferencePipeline.collect(ReferencePipeline.java:578)
	at org.projectnessie.versioned.impl.KeyList$IncrementalList.getKeysIter(KeyList.java:145)
	at org.projectnessie.versioned.impl.KeyList$IncrementalList.generateNewCheckpoint(KeyList.java:123)
	at org.projectnessie.versioned.impl.KeyList$IncrementalList.createCheckpointIfNeeded(KeyList.java:107)
	at org.projectnessie.versioned.impl.InternalL1.withCheckpointAsNecessary(InternalL1.java:68)
	at org.projectnessie.versioned.impl.InternalBranch.getUpdateState(InternalBranch.java:277)
	at org.projectnessie.versioned.impl.AbstractITTieredVersionStore.checkpointWithUnsavedL1(AbstractITTieredVersionStore.java:314)
```

Fixes #813