Add changeId to commit metadata.
This very simple change adds a `change id` to commit metadata.

This is meant to be distinct from the commit id in that it will never change even over rebases and merges. This is primarily used to uniquely identify an object in the store. The tuple (CommitMeta.changeId,Contents.uuid) is immutable for a commit and unique across the database.