Add profile "docker" to `:nessie-quarkus`
To build a container-image in `:nessie-quarkus` it's currently necessary to use the
`release` profile. But the `release` profile also signs the artifacts, so users have
to have set up all-the-things gpg and have the password handy.

Adding a new profile `docker` to just build the container image is more convenient.