Allow multiple `Test` tasks per Gradle project (#1221)
Previously, although the Gradle-Plugin code searched for all tasks that extend the `Test` task,
only the task named `test` was correctly configured. This patch fixes the issue and applies
the configuration (system properties) to all `Test` tasks.

Background: This is to support both "normal" unit tests and also integration-tests in Iceberg
using the `test` and a new `integrationTest`, the latter is to test Nessie against Iceberg
via Spark.