Handle `NessieConflictException` in Delta clients + handling of multiple-branches (#1249)
* Delta-client-tests: add build-resources from core package
* Handle `NessieConflictException` in Delta clients + handling of multiple-branches

The Delta clients can run into an unrecoverable (dead) state, when the reference changed in the mean time.
This commit adds a single retry in case the client throws a `NessieConflictException`.

The Delta-Client kept a single `Reference` that was initialized with the named-reference given
in the hadoop-config. However, if the configuration changed to switch to another Nessie branch,
that `Reference` was never refreshed and resulted into `NessieNotFoundException`s.

The change is to replace that single `Reference` field with a map of ref-name to `Reference`
to support multiple branches.

Also adds tests to verify that the Delta client works with multiple branches and the retry
when hitting a `NessieConflictExcepiton`. The new tests are disabled for Spark2/Delta0.6.