Fix release-notes-step during release-workflow (#1505)
In the GH WF step that creates the release-notes.md file, we need the commits
starting from the previous tag. However, the checkout-action provides "only" a
shallow clone (last commit), so there is no history.

The fix is to perform a "full" clone of the GH repo, which shouldn't be a big
deal, since the repo's history is linear.