Fix ITNativeRest issue
A native Quarkus application does not seem to have access to `quarkus.resteasy.path`, unlike a
non-native Quarkus application.

However, it's actually not necessary to read that configuration option, as it's possible to
retrieve that information implicitly from `UriInfo.getBaseUri()`.

So this change removes the `MultiTenantConfig` interface + implementation and adopts
`MultiTenantFilter`.