Add global-state to Contents
To make potentially conflicting changes (DDL or DML) to a table safe, Nessie needs to validate
certain global table attributes and act as a kind of "source of truth" for these attributes.
Those attributes are:
* last-column-id
* current-schema-id
* last-assigned-partition-id
* current-snapshot-id

Nessie must ensure that changes to the same table in different Nessie branches use globally
unique values for the above attributes.

"Global contents" information is meant to be tracked by Nessie for all tables across all
branches/tags.

"(Per branch) contents" contains information that is relevant to the Nessie branch/tag.

The approach here is to introduce a new schema-type hierarchy `GlobalContents` next to
`Contents` plus a new `Operation.PutGlobal`. A commit-operation must then pass two operations
for the same `ContentsKey`.
* a `Operation.Put` with the `Contents` object and
* a `Operation.PutGlobal` with the expected `GlobalContents` object and the updated
 `GlobalContents` object.

It feels that a new exception next to `NessieConflictException` is not necessary and Nessie
can raise/return a `NessieConflictExcepiton` in case the change from "expected global state"
to "proposed global state" cannot succeed.

Although we have technically two "contents" objects (`IcebergTable` + `IcebergSnapshot`),
we do not have to return both objects from the `ContentsApi`, because the only global
attribute (table-manifest-location) is present in both "contents" objects.