Remove 'paths-ignore' for PRs due to mismatch with 'required status checks to pass'
Problem: Users opening PRs that only update ignored files, such as a .md
file end up having PRs that can't be merged because we have `Require status checks to pass before merging`
enabled for the `main` branch.
This is a known GH issue as reported at https://github.community/t/feature-request-conditional-required-checks/16761.
There seem to be a few workarounds mentioned on the ticket, however we
don't want to invest time currently in fixing this on our side as it
seems cheaper to run CI for those "ignored" files.