TVS2 allow nested key-lists (#1763)
The current implementation aggregates all known content-`Key`s in `CommitLogEntry.keyList`, so a larger amount of content-keys would hit a database's hard (and also highly recommended) limit of a database-object/database-row.

This PR enhances the "embedded key list" `CommitLogEntry.keyList` with "nested" `KeyListEntity`s, so the total number of keys is rather "unlimited".

When a "complete-key-list" is constructed, the total size of both `CommitLogEntry` and `KeyListEntity` is limited to `DatabaseAdapterConfig.getMaxKeyListSize`. The "global default" for this value is unlimited (`Integer.MAX_VALUE`). Databases that do force a hard/highly-recommended limit must override `DatabaseAdapterConfig.getDefaultMaxKeyListSize`, to imply that as the default for the concrete database.

Current tests (not in this PR) exercise against a rather small value for max-key-list-size and pass for the current implementations (not in this PR) for in-memory, rocks-db, h2 + postgres.