API-versioning w/ HTTP headers - changes to nessie-http
Another apporach to add "versioning" to the (REST) API is to leverage the `Content-Type` + `Accept` headers,
which can be "selected" in JAX-RS via the `@Consumes` + `@Produces` annotations. I.e. a changed API endpoint
keeps its old implementation and gets another one using the same `@Path` and HTTP-method but with different
media-types.

This change just prepares the `HttpClient` + `HttpRequest` classes for this scenario.