Enhance commit log
Allows to optionally return additional information from the commit log.

The `LogResponse` type has been changed to contain a list of `LogEntry` elements, which contain a `CommitMeta` and, if requested, also the operations (`Put` or `Delete`) of each commit.

The Python CLI behaves as it currently does. The `nessie log` command got a new flag to retrieve and show the additional information.

* Adds new parameter `fetchAdditionalInfo` to commit-log REST endpoint.
* Changes return type of commit-log REST endpoint, returns a list of `LogEntry`
  that contains the `CommitMeta` plus the optionally fetched list of operations.
* Adds a `--extended` flag to Nessie CLI's `log` command to include the additional
  information. Behaves exactly as before without the `--extended` flag.