Add noop-BackwardsCompatibleVersionStore / prepare global-states PR
The global-states-PR will introduce a new operation type (`PutGlobal`) and change contents (make `IcebergTable` a
global-state + introduce `IcebergSnapshot`). While we have the jgit + in-memory + dynamo version-stores in the
code base, a "translation layer" is needed to map the global-state-ish operations + contents to those version-store
implementations.

This PR only introduces `BackwardsCompatibleVersionStore`, which will be fleshed out with the actual global-states-PR.