TVS/non-tx: Fix paging for global-states-log + add validation to concurrency tests
The intention of `List GlobalStateLogEntry.parents` is to track a list of parent-hashes and benefit from fetching
multiple keys at once when traversing the global-state-log.

However, `NonTransactionalDatabaseAdapter.writeGlobalCommit()` has a bug that actually hard-limited the number of
parent hashes to exactly 1. This is a preformance issue.

Also adds some validation to the concurrency-tests