Global-states: API changes (non-intrusive) (#1922)
* Global-states: API changes (non-intrusive)

This PR only adds API+model changes, but does **not** change any behavior. Actual behavior changes and
implementations for global-state are subject to a follow-up PR.

Introduces the model changes for global-state.

* `IcebergTable` gets a new attribute `snapshotId`. The already existing `metadataLocation` will represent
  the "global state" and `snapshotId` the "per-reference state`

Operations:

* `Put` updates the `IcebergTable` for a key and gets a new attribute `expectedValue`. When updating an
  Iceberg table, clients pass the "last known" version of `IcebergTable` as the `expectedValue`. The server
  side automatically extracts the global-state from that `expectedValue`.
* No put-global operation.

No other API changes.

The initial version for global-states looked similar to this one. An intermediate version of this change
did model the global-state and on-reference-state as separate object types - on-reference-states were
updated using `Put` and global-states updated using an added `PutGlobal` operation. The latter approach
distinguished the meaning of the individual objects more clearly, but was much more difficult to implement
and use.

* Fix docs