Trigger CI only for labels starting with the string `pr-`
Currently, CI runs are triggered for every addition or removal of any label.
This causes PRs created by dependabot to trigger 3 CI runs:
* one for the PR creation (type `opened`)
* one for the addition of the `dependencies` label (type `labeled`)
* one for the addition of the package-type label (type `labeled`)

It would be ideal to have a filter for the workflow-triggers `labeled` and
`unlabeled`, but that does not exist.

Using a separate workflow that only listens on `labeled` and `unlabeled`
event types does not really help, because such a workflow would have to
be able to list and rerun check-suites via the GH API, but the permissions
for `pull_request` workflows to not permit that. So a separate WF is not
an option.

Easiest for now is to add an
```
if: (github.event.action != 'labeled' && github.event.action != 'unlabeled') || startsWith(github.event.label.name, 'pr-')
```
to the relevant jobs in `pull-request.yml` so that these jobs only run when
it's not a `labeled` or `unlabeled` type or if the label name starts with
the string `pr-`.

The disadvantate is that when adding or removing labels that do _not_
start with the string `pr-`, the jobs gets triggered but is not run and
therefore appears as `Skipped` in the list of PR checks and the "real"
PR-CI-run can only be inspected by clicking on the `Details` link and
the check appears as **_passed_**, which is NOT CORRECT.