Split PR workflow into three
Currently, CI runs are triggered for every addition or removal of any label.
This causes PRs created by dependabot to trigger 3 CI runs:
* one for the PR creation (type `opened`)
* one for the addition of the `dependencies` label (type `labeled`)
* one for the addition of the package-type label (type `labeled`)

It would be ideal to have a filter for the workflow-triggers `labeled` and
`unlabeled`, but that does not exist.

It feels possible that splitting the `pull-request.yml` workflow into
three workflows could solve the problem. The current one is run for
the types `opened`, `synchronize`, `reopened`. The two new ones, one
for "native tests" and one for "jackson tests", get additionally triggered
for the type `labeled`. `if:` conditions on the jobs in the new workflows
ensure that the jobs only run when the appropriate label is present.