TVS: add microbenchmark base (#2084)
Adds a first JMH based microbenchmark against database-adapters.
Microbenchmarks are not run during CI though.

This PR also introduces a change to `TestConnectionProviderSource` so a suitable instance can be discovered and configured without adding database-adapter-specific code to each microbenchmark. That code can later be reused for the Gatling based tests. Goal of that change is that is should be sufficient to provide a database-adapter name plus system-properties to configure the connection-provider and database-adapter instances.