Separate configuration for # of parents in GlobalLogEntry
The implementation is currently reusing the number of parents per CommitLogEntry for
the number of parents in a GlobalLogEntry. The payload sizes for both types are very
different and it makes sense to have different configuration options and different
defaults for both. CommitLogEntry can be assumed to be bigger than GlobalLogEntry.
It also makes sense to have more parents per GlobalLogEntry, because we do not (yet)
aggregate the global-contents-values within the global-log.