Re-set commit meta on transplant/merge
Fixes [831](https://github.com/projectnessie/nessie/issues/831)

The patch attempts following changes on the merge/transplant commits -
* Committer cahnged to user principal, empty string when user is anonymous
* Commit time set to current time
* A property in the commit meta, with key [merge/transplant]_ref, and the value incoming branch name.

Approach -
The user principal is available in the TreeResource. This is required for setting the committer. However, copying/shifting commits are done in the layers below (DatabaseAdapter or VersionStore). Hence, approach prepares a lambda function, to translate commit metadata, and set properties.
The version store allows generic metadata object, and has access to serializer class (which is not available in the DBAdapter). Hence, version store bears responsibility of supplying the same on the function passed for metadata reset.

NOTE: Ignored the changes for the merge in JGit implementation, because it applies references of the commit rather than doing a copy. Since this is old version store implementation, it is ok to leave it for now.