Use default reference as base in CLI branch/tag commands
CLI has a config setting for a local default branch. This branch
(as documented) should be used as the "base" reference in branch/tag
commands unless the user explicitly specifies the base reference.

Moreover, the HEAD of the base reference should be used as the hash
for creating/assigning branches and tags unless the user explicitly
specifies another hash.

Fixes #2271
Fixes #2276