Explicitly define behavior of multiple operations for the same key in a commit
While thinking about how Nessie could actually squash multiple commits, I realized
that this can result in multiple commit-operations for the same content-key.

While commit-operations are ordered on the (REST) API level (list of `Operation`s),
database adapters split those into separate collections of delete and put operations,
which makes it impossible to determine the original order of the squashed operations.

The Nessie spec does not define the behavior of a Nessie commit with multiple
operations for the same key in one commit.

The spec now clearly states that there must be no more than operation (put / delete)
for each content key.

This commit updates the spec and also adds tests on the `DatabaseAdapter` and
`VersionStore` levels and includes related fixes to `AbstractDatabaseAdapter`
and `PersistVersionStore`.