Client compatibility updates
The commit updates the client to be compatible with the latest code. There's a cicrular dependency between iceberg and nessie primarily for integration testing and is yet to be addressed. While this change is still in a draft stage, it can be used for development needs as suited.

TODO: Resolve a way to handle circular dependency