Remove premature `prehash=true` in most cases
A lot of the immutables eagerly precalculate the hashCode, which is in most cases
just unnecessary and wastes both CPU and heap. Even most of the types that are e.g.
used as a map-key or added to a set are simple enough (e.g. contain a single string),
that it's not worth to maintain a separate field for those.