Add `-Dquickly` for faster local builds
A `./mvnw -Dquickly` performs a `clean install` without running tests. To speed things up,
add e.g. `-T1.5C` or use `mvnd`.

The presence of the `quickly` property omits a couple of rarely touched modules:
* ui
* perftest
* lambda

Also, it does not build the following modules to speed up the build (omit Scala):
* deltalake
* spark-extensions

Also fixes dependencies for the `:nessie-ui` module when running parallel Maven builds,
as the `:nessie-ui` module did not properly "wait for" `:nessie-model` with `-T...` or
`mvnd`.