Fix Swagger API example schema for getContents()
The swagger api example is partially fixed with https://github.com/projectnessie/nessie/pull/2545. However, the example wasn't executing because the schema was set to that of ContentsKey; resulting in a - not a valid json error"

This change attempts to
a. Use String type as Key @schema to allow execution from swagger.
b. Reset example key as "example.key" as it aligns with other examples in there.