Enhance `DatabaseAdapter.namedRefs()` (#2726)
Goal of this PR is to optionally expose more information about named
references. The caller decides which kind of information shall be
retrieved. The default behavior is unchanged and only returns a
collection of reference names + type + HEAD hash.

`DatabaseAdapter.namedRefs()` gets a new argument `GetNamedRefsParams`,
that allows to request:

* exclusion of tags in the result
* exclusion of branches in the result
* retrieving the commit-meta of the named reference's HEAD
* retrieving the common-ancestor of the named reference and the default branch
* retrieving the number of commits behind + ahead relative to the common-ancestor

The type describing each named reference in the result of
`DatabaseAdapter.namedRefs()` has been changed to
`ReferenceInfo<METADATA>` and allows returning the selected information.

The new types `ReferenceInfo` and  `GetNamedRefsParams` are defined
in `:nessie-versioned-spi` and are not meant to be exposed via the
REST API. Updating the REST API is work for a separate PR.