Refactor `getNamedRef` to take a plain `String` instead of `NamedRef` (#2766)
This eliminates two unnecessary roundtrips:
* removes one call to `VersionStore.toRef()` in `TreeApiImpl.getReferenceByName()`
* Eliminiates the unnecessary DB roundtrip in `PersistVersionStore.toRef()` when
  the reference parameter is not a branch name

This change slightly changes the behavior of `VersionStore.getNamedRef()` as it allows to query a named reference only by its name and return either the matching branch or tag. However, it's still possible to restrict `getNamedRef` to only search for/return a specific type (branch/tag) via the parameter taking the `GetNamedRefsParams`.

Note: if the caller only needs the hash of a reference's HEAD, it's better to use `VersionStore.hashOnReference()` with an empty optional.