Track total number of commits (#2767)
* Track total number of commits

Allows retrieval of the total number of commits on a reference.

Via REST: in the `ReferenceMetadata`, if `fetchAdditionalInfo` is true.
Via VersionStore/DatabaseAdapter: if `retrieveCommitMetaForHead` is true.

The total number of commits is tracked as a field in every commit. The
first commit since "beginning of time" starts at `1`, following commits
"increment" this number. Since the commit history is immutable, the
number of commits since "beginning of time" is immutable as well.

* review

* only add info if requested