Let Mongo tests via flapdoodle use Unix sockets
Using anything relying on TCP ports is actually bad:
* Using the local port from a previous `new ServerSocket(0).getLocalPort()` is racy
  (by definition)
* Using a TCP port from another PRNG is bad
* Using a fixed port is bad (you can't run tests in parallel)
* Telling mongod to use "port 0" and parse the log output is also error prone

But it's possible to let Mongod listen to a Unix domain socket, which is a file,
and we can pretty much guarantee (at least way better than the above TCP port
approaches) that the FS path for the Unix domain socket is "ours".

The only downside is that tests require `com.github.jnr:jnr-unixsocket`, but it's
for tests only, so feels acceptable.