Track Nessie database repo state/version
With the current work on implementing the ref-log it becomes kinda obvious that repository migration on the database level need to be handled.

The idea in this change is to have a description about the repository that provides these attributes:
* `repo_version` an integer version describing the repository, but not the database layout
* `properties` as a generic key-value bag describing the repository configuration

Behavior:
* if no repository-description exists, assume version `0` and an empty properties

Version `0` would be what we have now (including the ref-log).

Fixes #3009