Ensure stable order of references in GlobalStatePointer (#3111)
As pointed out in the review of #3028, Java's `Map` does not ensure a stable order of elements.
This was definitely a problem in that PR.

Nessie's `GlobalStatePointer` also uses a `Map` for the named references. Although the
design of the code ensures that the order of elements in `GlobalStatePointer.namedReferences`
is stable, it feels safer to refactor that property to a list.

The "order" of the list is: most recently updated named references appear at the beginning of
that list. This change brings more advantages beside the "defined order or elements":
* Most recently updated references appear before "stale" references.
* Allows to "spill out" stale references to a separate table, if that becomes necessary, i.e.
  when there are too many references, the GSP becomes big and databases suffer from that size.
* (Probably neglectible:) a bit less (de)serialization effort as the map on top of the protobuf
  list is gone.