Update committer to current user during merge+transplant (#3122)
This change updates the "committer" to the current user in the commits created by merge + transplant.

Tests to validate the behavior "with security" (i.e. when there's a `SecurityContext` available) are included in `AbstractTestJerseyRest`. Updated the `NessieJaxRsExtension` to provide a `@NessieSecurityContext` test parameter to provide the `SecurityContext` from tests.

Fixes #831