Just make it work ... not nice though
Splitting `AbstractTestRest` into separate classes and including those
via `@Nested` with plain JUnit 5 works fine. But it does not work as
a `@QuarkusTest`.

Using `@Nested` test classes like this:
```
public void MyTest {
  @Nested
  public class Group {
    @Test
    void foo() {}
  }
}
```
does not work, because Quarkus considers the non-static inner class
`Group` as a separate test class and fails horribly, mostly complaining
that `...MyTest$Group` is not in one of the expected class folders.

Even worse is that the whole Quarkus machinery is started for each
nested test class and not just for `MyTest`.

As an alternative it felt doable to refactor the former inner classes
to interfaces (so test methods become default methods), but that did
not work as well due to an NPE, when
`QuarkusSecurityTestExtension.beforeEach` tries to find the `@TestSecurity`
annotation - it just asserts whether the current `Class` is `!= Object.class`,
but `Class.getSuperclass()` returns `null` for interfaces - hence the
NPE there.

The "solution" here is to keep the separate classes but let each
extend another - so different groups of tests are kept in a single class.