Update Maven+Gradle plugins to launch the Nessie-Server via `java -jar` (#1344)
Before this PR, the apprunner-gradle-plugin uses Quarkus mechanisms to resolve artifacts, setup the configuration, etc. - this reuses functionality from the apprunner-maven-plugin. Since #1245 (new in Nessie 0.6.1), the Nessie-Server is built with Java 11.

The idea is to launch the Nessie server as a separate process using the uber-jar, which is published to Maven Central since Nessie 0.6.1.

Removing all the Quarkus-bootstrap code from the plugin simplifies the plugin code a lot and removes the amount of required dependencies.

There are 2 "official" and 1 "private" way to specify the Nessie-Quarkus jar to run:
1. In Maven by specifying the coordinates via the `appArtifactId` config options, in Gradle by declaring the dependency via the `nessieQuarkusServer` configuration).
2. In Maven or Gradle by providing the path to the jar via the `executableJar` config option.
3. Only in Maven and only for Nessie-source-tree builds: neither 1 nor 2, the Maven plugin figures out the path on its own.

A new module `nessie-apprunner-common` contains the underlying code to find the JVM via environment variables or system properties. Then Maven tries to find a matching JDK via the Maven toolkits API and falls back to the `JavaVM` implementation. The Gradle plugin always uses the `JavaVM` implementation (Gradle 5.4 doesn't have that functionality, but newer Gradle version do). That module also contains code to start the Java process running the Nessie-Server-Quarkus-jar, capture the http listen URL and provide both the full URL and port via project properties to associated test tasks.

The overall implementation makes it much easier to reason about system properties and environment variables passed to the executable Nessie-Quarkus-jar. With the Quarkus-bootstrap code, it was hard to reason about and debug issues with properties passed to Quarkus and/or Nessie.

The Maven plugin also got some integration tests.

Fixes #1333