Replace StringStoreWorker with something more expressive
Introduces "real" types for contents + commit-meta that shadow the
behavior of content types without global state and with global state.
These test content types also always have a content-id, as the
"real" types in `:nessie-model` do.

The actual changes are in `:nessie-versioned-tests` in the package
`org.projectnessie.versioned.testworker`.

All other changes are rather "boring".

This is a prerequisite for storing Iceberg table/view metadata
in Nessie.