Support Namespaces
This introduces a new `NamespaceApi` that allows creating/retrieving/deleting a single Namespace and also retrieving all Namespaces with a given prefix.

A Namespace is now a subtype of `Content` to keep changes to a minimum at the storage level.

Given that users currently don't explicitly create a Namespace when adding a new table 'a.b.c.table', the `NamespaceApi` is designed in a way where it first needs to evaluate:
* whether an explicitly created Namespace exists: An explicitly created Namespace is created via `NamespaceApi#createNamespace(NamespaceParams)`
* whether an implicitly created Namespace exists: An implicitly created namespace generally occurs when adding a table 'a.b.c.table' where the 'a.b.c' part represents the namespace without

This feature also adds 3 new Exceptions: `NessieNamespaceAlreadyExistsException` / `NessieNamespaceNotEmptyException` / `NessieNamespaceNotFoundException`

Note that this currently does not support Namespace properties, as those will be handled separately.